+++ gui/layout_sizing_fuzz_test.go
@@ -307,7 +307,7 @@ func FuzzLayoutSizingWithMix(f *testing.F) {
 func shapeTypeFromUint(v uint32) shapeType {
 	types := []shapeType{
 		shapeRectangle, shapeCircle, shapeText, shapeImage,
-		shapeSVG, shapeRTF, shapeNone,
+		shapeSVG, shapeRTF, shapeTermGrid, shapeNone,
 	}
 	return types[v%uint32(len(types))]
 }